Alternatively lookup vars either from yaml or env
This patch add a get method used to abstract the lookup between
the env and the yaml config file.
I also removed the Exception in case of the yaml file is not found
because we don't care about a config file when using only env vars.
You could now either have a config.yaml with some modes in it or
just use the env vars.
The env vars have a greater precedence on everything.

You can also specify an override_mode in _getValue() in case you want
to fetch the var from another mode specified in the config.yaml file

to use it: cls._getValue("OSX_SLAVE_NAME", override_mode="test")