Check all processes when determining connection info on Windows
This has the consequence of making the `pid` entry in the connection misleading, but the more important thing is the user for now. We can modify this to include all pids if we need to, but right now the higher priority is determining the root cause of some multi-tenant issues.