Check all processes for user ownership in connection info on Windows (#118)
* Check all processes when determining connection info on Windows

This has the consequence of making the `pid` entry in the connection misleading, but the more important thing is the user for now. We can modify this to include all pids if we need to, but right now the higher priority is determining the root cause of some multi-tenant issues.

* Use an array and `in` for user inclusion instead of a set

user info structures don't define __hash__, and I don't want to get in to that.